fix(dtype): use dtype of inputs

+++ torchimize/functions/parallel/lma_fun_parallel.py
@@ -95,7 +95,7 @@ def lsq_lma_parallel(
         p, f, g, H = newton_step_parallel(p, fun, jac_fun, wvec)
         D = lm_dg_step(H, D)
         Hu = H+u[:, None, None]*D
-        h = -torch.linalg.lstsq(Hu.double(), g.double(), rcond=None, driver=None)[0]
+        h = -torch.linalg.lstsq(Hu.double(), g.double(), rcond=None, driver=None)[0].to(dtype=p.dtype)
         f_h = fun(p+h)
         rho_nom = torch.einsum('bcp,bci->bc', f, f).sum(1) - torch.einsum('bcp,bci->bc', f_h, f_h).sum(1)
         rho_denom = torch.einsum('bnp,bni->bi', h[..., None], (u[:, None]*h-g)[..., None])[..., 0]
